refactor(pricefeed): move MockSource to pricefeedtest subpackage
The price feed mock was exported from the production pricefeed package,
so it could be imported (accidentally or otherwise) by non-test code.
Moving it to a sibling pricefeedtest package — following the stdlib
httptest/iotest convention — keeps it reachable from every test that
needs it while preventing accidental use from production binaries.

aggregator_test.go switches to an external _test package (pricefeed_test)
because the mock now lives outside pricefeed.
